FilingCabinet and working and a remove() test
Added a remove() test method to ChunkArrayTest, which will probably
need to be edited to check for the chunk removal eventually, or at
least somehow test that.

Filing Cabinet is getting a new make file method, so that you can
create blank documents. That way, one can add stuff in to it using the
print writer method, instead of just making specified files.