Error Logs and Fixed Chunk Demo
ChunkArray is now working as it should for ChunkDemo. The chunk removal
method was overkill in the time it took, and the threading in the front
end couldn’t deal with the asynchronous access there for some reason. I
think Chunk Demo should do a test with a growing circle to test the
limits of the program. Also, please check out what I did, because its
pretty weird. I limited the ArrayList.size() method by 1 on several
occasions, which means we either don’t know what it does or we screwed
up.