Built the coordinate class, and winning conditions
I made a class to be applied later, which takes objects as Coordinates,
and then can be used abstractly by players by concretely by the system.
Also, I implemented the state winning methods in the TicTacGrow class,
which not only gave me more things to do, but allowed things to be
executed again.